cache workspace results to rate limit slack integrations
